fix(api): split up api utils to prevent frontend import of mongoose package
