+++ packages/api/src/user/utils/core.ts
@@ -1,0 +1,5 @@
+export const shortenEthAddress = (address: string): string => {
+  return `${address.substring(0, 6)}...${address.substring(
+    address.length - 4
+  )}`;
+};
